requirements.txt
# AI Shorts Factory - Python依存パッケージ
# Python 3.11+

# === 共通 ===
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
aiofiles==24.1.0
tenacity==9.0.0

# === 台本生成（Track A） ===
huggingface-hub==0.27.1

# === 音声生成（Track B） ===
# VOICEVOX Engine API（HTTP経由で接続）
# 追加パッケージ不要

# === 画像生成（Track C） ===
gradio-client==1.5.3
Pillow==10.4.0

# === 動画素材取得（Track D） ===
# Pexels API（httpx使用）
# 追加パッケージ不要

# === 動画合成（Track E） ===
moviepy==2.1.2
openai-whisper==20240930
numpy==1.26.4

# === YouTube投稿（Track F） ===
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# === テスト ===
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# === 開発ツール ===
ruff==0.9.2
mypy==1.14.1
